Allow ExecStartPost and ExecStopPre to fail
There's two things required to make that happen. First, the Execs are
now marked with "-" to instruct systemd to ignore non-zero returncodes
of said commands. Second, the commands are now run under the
supervision of the 'timeout' program to not run into systemd timeouts.

This change helps avoiding systemd hard-killing the VM (fixes #9)
when a command fails or when it would take to long to finish within
given timeouts.